Move completed todo archive warning to todos context
