+++ loopx/control_plane/work_items/project_asset.py
@@ -4,7 +4,6 @@ import re
 from typing import Any, Callable
 
 
-DEFAULT_MAX_ACTIVE_DONE_TODOS_BEFORE_ARCHIVE = 12
 DEFAULT_MONITOR_SIGNAL_WAITING_ON = "monitor_signal"
 DEFAULT_MONITOR_DISPLAY_STOP_CONDITION = (
     "stop until a material monitor transition, regression, or concrete blocker appears"
@@ -43,37 +42,6 @@ def project_asset_public_safe_compact_text(value: Any, *, limit: int = 220) -> s
     if LOCAL_PATH_SURFACE_PATTERN.search(text) or SECRET_LIKE_SURFACE_PATTERN.search(text):
         return None
     return text
-
-
-def completed_todo_archive_warning(
-    agent_todos: dict[str, Any] | None,
-    *,
-    max_active_done_todos: int = DEFAULT_MAX_ACTIVE_DONE_TODOS_BEFORE_ARCHIVE,
-) -> dict[str, Any] | None:
-    if not isinstance(agent_todos, dict):
-        return None
-    try:
-        done_count = int(agent_todos.get("done_count") or 0)
-    except (TypeError, ValueError):
-        done_count = 0
-    if done_count <= max_active_done_todos:
-        return None
-    try:
-        open_count = int(agent_todos.get("open_count") or 0)
-    except (TypeError, ValueError):
-        open_count = 0
-    return {
-        "kind": "completed_agent_todo_archive_required",
-        "requires_archive": True,
-        "archive_section": "Completed Work Archive",
-        "active_done_count": done_count,
-        "active_open_count": open_count,
-        "max_active_done_count": max_active_done_todos,
-        "recommended_action": (
-            "move older completed Agent Todo entries into a dedicated Completed Work Archive "
-            "until the active Agent Todo section keeps only current open work and a small recent-done tail"
-        ),
-    }
 
 
 def project_asset_owner(
